Adding a Questlog / interactions with NPCs
Bubbles also now have a ButtonList. This is an invisible variable that will be pushed onto the player, so they can't move during an interaction. NPC's have quests, that are just bubbles. When a player presses 'x', provided they are interacting with an NPC, they open up the quest bubble.