Finished "Talk" feature of shop
Full shop funcitonality is now implemented. If you want to change the inventory, you create an ArrayList of objects and use setInventory. Sell always works the same, as it is contingent on the Player's inventory, which will be updated throughout the game. The dialogue for "Talk" can be altered by using setDialogue(..), and passing a 2D array of Strings as the parameter. Each entry in this 2D array will store an array of Strings, where the first element holds the title of the interaction, and the subsequent Strings hold what the shopkeeper will say, with each distinct element representing another slide (the player will have to press 'x' to advance the dialogue between slides). Leave is also always the same for each shop.