Added portals
Room now has an ArrayList of portals. A portal has an image, two PVectors representing position and dimensions, and an associated Room that the player teleports to upon entering it. Also, a transition effect was added in Camera, through a new method "display". The transition effect can be executed via the panTo(Room) method.